static/tests/samesite/iframe.html

<!DOCTYPE html>
<script src="/tests/resources/testharness.js"></script>
<script src="/tests/resources/testharnessreport.js"></script>
<script src="/tests/resources/cookie-helper.js"></script>
<body>
<script>
  var HTTP_COOKIE = "cookie_via_http";

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => embed_cookie_posting_iframe(ORIGIN))
      .then(cookies => {
        assert_cookie(ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Same-host frames are laxly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => embed_cookie_posting_iframe(ORIGIN))
      .then(cookies => {
        assert_cookie(ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Same-host frames are strictly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(SUBDOMAIN_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => embed_cookie_posting_iframe(SUBDOMAIN_ORIGIN))
      .then(cookies => {
        assert_cookie(SUBDOMAIN_ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Subdomain frames are laxly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(SUBDOMAIN_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => embed_cookie_posting_iframe(SUBDOMAIN_ORIGIN))
      .then(cookies => {
        assert_cookie(SUBDOMAIN_ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Subdomain frames are strictly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(CROSS_SITE_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => embed_cookie_posting_iframe(CROSS_SITE_ORIGIN))
      .then(cookies => {
        assert_cookie(CROSS_SITE_ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, false);
      });
  }, "Cross-site frames are not laxly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(CROSS_SITE_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => embed_cookie_posting_iframe(CROSS_SITE_ORIGIN))
      .then(cookies => {
        assert_cookie(CROSS_SITE_ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, false);
      });
  }, "Cross-site frames are not strictly same-site.");

  //
  // Redirected variants of the above:
  //
  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => embed_cookie_posting_iframe(redirectTo(ORIGIN, ORIGIN)))
      .then(cookies => {
        assert_cookie(ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Same-host frames redirected to same-host are strictly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(SUBDOMAIN_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => embed_cookie_posting_iframe(redirectTo(ORIGIN, SUBDOMAIN_ORIGIN)))
      .then(cookies => {
        assert_cookie(SUBDOMAIN_ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Same-host frames redirected to subdomain are strictly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => embed_cookie_posting_iframe(redirectTo(SUBDOMAIN_ORIGIN, ORIGIN)))
      .then(cookies => {
        assert_cookie(ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "Subdomain frames redirected to same-host are strictly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(CROSS_SITE_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => embed_cookie_posting_iframe(redirectTo(ORIGIN, CROSS_SITE_ORIGIN)))
      .then(cookies => {
        assert_cookie(CROSS_SITE_ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, false);
      });
  }, "Same-host frames redirected to cross-site are not same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(CROSS_SITE_ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => embed_cookie_posting_iframe(redirectTo(SUBDOMAIN_ORIGIN, CROSS_SITE_ORIGIN)))
      .then(cookies => {
        assert_cookie(CROSS_SITE_ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, false);
      });
  }, "Subdomain frames redirected to cross-site are not same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Lax")
      .then(_ => embed_cookie_posting_iframe(redirectTo(CROSS_SITE_ORIGIN, ORIGIN)))
      .then(cookies => {
        assert_cookie(ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "(TODO?) Cross-site frames redirected to same-host are laxly same-site.");

  promise_test(t => {
    var COOKIE_VALUE = "" + Math.random();
    return create_cookie(ORIGIN, HTTP_COOKIE, COOKIE_VALUE, "SameSite=Strict")
      .then(_ => embed_cookie_posting_iframe(redirectTo(CROSS_SITE_ORIGIN, ORIGIN)))
      .then(cookies => {
        assert_cookie(ORIGIN, cookies, HTTP_COOKIE, COOKIE_VALUE, true);
      });
  }, "(TODO?) Cross-site frames redirected to same-host are strictly same-site.");
</script>
